feat: created and updated plans

+++ web/src/pages/Student/ListStudents/index.js
@@ -1,10 +1,12 @@
 import React, { useState, useEffect } from 'react';
-import { MdAdd } from 'react-icons/md';
+import { MdAdd, MdSearch } from 'react-icons/md';
+import { Link } from 'react-router-dom';
+import { toast } from 'react-toastify';
 
+import Button from '~/components/Button';
 import api from '~/services/api';
-import history from '~/services/history';
 
-import { Container, Content, Wrapper, Grid, GridButton } from './styles';
+import { Container, Wrapper, Grid, GridButton, Search } from './styles';
 
 export default function ListStudents() {
   const [students, setStudents] = useState([]);
@@ -19,16 +21,22 @@ export default function ListStudents() {
     loadStudents();
   }, [students]);
 
-  function handleSubmit() {
-    history.push('/created/students');
-  }
-
-  function handleUpdateStudent(id) {
-    history.push(`/students/${id}`);
-  }
+  useEffect(() => {
+    document.title = 'Gympoint | Alunos';
+  }, []);
 
   async function handleDeleteStudent(id) {
-    await api.delete(`/students/${id}`);
+    const result = window.confirm(
+      'Você realmente quer excluir este estudante?'
+    );
+    if (result) {
+      try {
+        await api.delete(`/students/${id}`);
+        toast.success('Studante deleted with success!');
+      } catch (err) {
+        toast.error('Studante deleted failure!');
+      }
+    }
   }
 
   return (
@@ -36,42 +44,45 @@ export default function ListStudents() {
       <Wrapper>
         <h1>Gerenciando alunos</h1>
         <div>
-          <button type="submit" onClick={handleSubmit}>
-            <MdAdd size={24} color="#fff" />
-            CADASTRAR
-          </button>
-          <input
-            type="text"
-            placeholder="Buscar aluno..."
-            onChange={setStudent}
-            value={student}
-          />
+          <Link to="/dashboard/students/created">
+            <Button type="submit">
+              <MdAdd size={20} color="#fff" />
+              CADASTRAR
+            </Button>
+          </Link>
+          <Search>
+            <MdSearch size={20} color="#ccc" />
+            <input
+              type="text"
+              placeholder="Buscar aluno..."
+              onChange={e => setStudent(e.target.value)}
+              value={student}
+            />
+          </Search>
         </div>
       </Wrapper>
-      <Content>
-        <Grid>
-          <li>
-            <strong>NAME</strong>
-            <strong>E-MAIL</strong>
-            <strong>IDADE</strong>
+      <Grid>
+        <li>
+          <strong>NOME</strong>
+          <strong>E-MAIL</strong>
+          <strong>IDADE</strong>
+        </li>
+        {students.map(s => (
+          <li key={s.id}>
+            <p>{s.name}</p>
+            <p>{s.email}</p>
+            <p>{s.age}</p>
+            <GridButton>
+              <Link to={`/dashboard/students/updated/${s.id}`}>
+                <button type="submit">EDITAR</button>
+              </Link>
+              <button type="submit" onClick={() => handleDeleteStudent(s.id)}>
+                EXCLUIR
+              </button>
+            </GridButton>
           </li>
-          {students.map(s => (
-            <li key={s.id}>
-              <p>{s.name}</p>
-              <p>{s.email}</p>
-              <p>{s.age}</p>
-              <GridButton>
-                <button type="submit" onClick={() => handleUpdateStudent(s.id)}>
-                  EDITAR
-                </button>
-                <button type="submit" onClick={() => handleDeleteStudent(s.id)}>
-                  EXCLUIR
-                </button>
-              </GridButton>
-            </li>
-          ))}
-        </Grid>
-      </Content>
+        ))}
+      </Grid>
     </Container>
   );
 }
